Create new test module for force_locale
