+++ tests/test_date_formatting.py
@@ -2,7 +2,6 @@
 from __future__ import with_statement
 
 from datetime import datetime, timedelta
-from threading import Semaphore, Thread
 
 import flask
 
@@ -107,60 +106,3 @@ def test_refreshing():
         app.config['BABEL_DEFAULT_TIMEZONE'] = 'Europe/Vienna'
         babel.refresh()
         assert babel.format_datetime(d) == 'Apr 12, 2010, 3:46:00 PM'
-
-
-def test_force_locale():
-    app = flask.Flask(__name__)
-    b = babel.Babel(app)
-
-    @b.localeselector
-    def select_locale():
-        return 'de_DE'
-
-    with app.test_request_context():
-        assert str(babel.get_locale()) == 'de_DE'
-        with babel.force_locale('en_US'):
-            assert str(babel.get_locale()) == 'en_US'
-        assert str(babel.get_locale()) == 'de_DE'
-
-
-def test_force_locale_with_threading():
-    app = flask.Flask(__name__)
-    b = babel.Babel(app)
-
-    @b.localeselector
-    def select_locale():
-        return 'de_DE'
-
-    semaphore = Semaphore(value=0)
-
-    def first_request():
-        with app.test_request_context():
-            with babel.force_locale('en_US'):
-                assert str(babel.get_locale()) == 'en_US'
-                semaphore.acquire()
-
-    thread = Thread(target=first_request)
-    thread.start()
-
-    try:
-        with app.test_request_context():
-            assert str(babel.get_locale()) == 'de_DE'
-    finally:
-        semaphore.release()
-        thread.join()
-
-
-def test_refresh_during_force_locale():
-    app = flask.Flask(__name__)
-    b = babel.Babel(app)
-
-    @b.localeselector
-    def select_locale():
-        return 'de_DE'
-
-    with app.test_request_context():
-        with babel.force_locale('en_US'):
-            assert str(babel.get_locale()) == 'en_US'
-            babel.refresh()
-            assert str(babel.get_locale()) == 'en_US'
